+++ frontend/src/components/index.ts
@@ -12,6 +12,6 @@ export { default as EventForm } from './events/EventForm.vue';
 export { default as EventSubNav } from './events/EventSubNav.vue';
 
 // User Form Components
-export { default as LoginForm } from './user-forms/LoginForm.vue';
-export { default as PasswordChangeForm } from './user-forms/PasswordChangeForm.vue';
-export { default as UserForm } from './user-forms/UserForm.vue';
+export { default as LoginForm } from './auth/LoginForm.vue';
+export { default as PasswordChangeForm } from './auth/PasswordChangeForm.vue';
+export { default as UserForm } from './auth/UserForm.vue';
